+++ packages/cad-simple-viewer/src/editor/global/index.ts
@@ -1,2 +1,3 @@
 export * from './AcEdGlobalFunc'
+export * from './AcEdConditionWaiter'
 export * from './eventBus'