+++ packages/cad-simple-viewer/src/view/AcTrView2d.ts
@@ -26,6 +26,7 @@ import { AcApSettingManager } from '../app'
 import {
   AcEdBaseView,
   AcEdCalculateSizeCallback,
+  AcEdConditionWaiter,
   AcEdCorsorType,
   AcEdViewMode,
   eventBus
@@ -94,6 +95,8 @@ export class AcTrView2d extends AcEdBaseView {
   private _stats: Stats
   /** Map of missing raster images during rendering */
   private _missedImages: Map<AcDbObjectId, string>
+  /** The number of entities waiting for processing */
+  private _numOfEntitiesToProcess: number
 
   /**
    * Creates a new 2D CAD viewer instance.
@@ -169,6 +172,7 @@ export class AcTrView2d extends AcEdBaseView {
     this.onWindowResize()
     this.animate()
     this._isDirty = true
+    this._numOfEntitiesToProcess = 0
   }
 
   /**
@@ -360,12 +364,20 @@ export class AcTrView2d extends AcEdBaseView {
   /**
    * @inheritdoc
    */
-  zoomToFit() {
-    if (this._scene.box) {
-      const box = AcTrGeometryUtil.threeBox3dToGeBox2d(this._scene.box)
-      this.zoomTo(box)
-      this._isDirty = true
-    }
+  zoomToFit(timeout: number = 0) {
+    const waiter = new AcEdConditionWaiter(
+      () => this._numOfEntitiesToProcess <= 0,
+      () => {
+        if (this._scene.box) {
+          const box = AcTrGeometryUtil.threeBox3dToGeBox2d(this._scene.box)
+          this.zoomTo(box)
+          this._isDirty = true
+        }
+      },
+      300, // check every 200 ms
+      timeout
+    )
+    waiter.start()
   }
 
   /**
@@ -448,6 +460,7 @@ export class AcTrView2d extends AcEdBaseView {
    */
   addEntity(entity: AcDbEntity | AcDbEntity[]) {
     const entities = Array.isArray(entity) ? entity : [entity]
+    this._numOfEntitiesToProcess += entities.length
     setTimeout(async () => {
       await this.batchConvert(entities)
     })
@@ -623,17 +636,23 @@ export class AcTrView2d extends AcEdBaseView {
           !(threeEntity as AcTrGroup).isOnTheSameLayer
         ) {
           this.handleGroup(threeEntity as AcTrGroup)
+          this.decreaseNumOfEntitiesToProcess()
         } else {
           const isExtendBbox = !(
             entity instanceof AcDbRay || entity instanceof AcDbXline
           )
 
-          await threeEntity.draw().then(() => {
-            this._scene.addEntity(threeEntity, isExtendBbox)
-            // Release memory occupied by this entity
-            threeEntity.dispose()
-            this._isDirty = true
-          })
+          await threeEntity
+            .draw()
+            .then(() => {
+              this._scene.addEntity(threeEntity, isExtendBbox)
+              // Release memory occupied by this entity
+              threeEntity.dispose()
+              this._isDirty = true
+            })
+            .finally(() => {
+              this.decreaseNumOfEntitiesToProcess()
+            })
         }
 
         if (entity instanceof AcDbViewport) {
@@ -655,6 +674,8 @@ export class AcTrView2d extends AcEdBaseView {
           const fileName = entity.imageFileName
           if (fileName) this._missedImages.set(entity.objectId, fileName)
         }
+      } else {
+        this.decreaseNumOfEntitiesToProcess()
       }
     }
   }
@@ -688,5 +709,15 @@ export class AcTrView2d extends AcEdBaseView {
       entity.dispose()
     })
     this._isDirty = true
+  }
+
+  private decreaseNumOfEntitiesToProcess() {
+    this._numOfEntitiesToProcess--
+    if (this._numOfEntitiesToProcess < 0) {
+      this._numOfEntitiesToProcess = 0
+      console.warn(
+        'Something wrong! The number of entities to process should not be less than 0.'
+      )
+    }
   }
 }