+++ packages/cad-simple-viewer-example/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mlightcad/cad-simple-viewer-example
 
+## 1.1.8
+
+### Patch Changes
+
+- feat: fix bug on zoomToFit
+- Updated dependencies
+  - @mlightcad/cad-simple-viewer@1.1.8
+
 ## 1.1.7
 
 ### Patch Changes