+++ packages/three-renderer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mlightcad/three-renderer
 
+## 1.1.8
+
+### Patch Changes
+
+- feat: fix bug on zoomToFit
+
 ## 1.1.7
 
 ### Patch Changes